+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/views/CodePreferencesPane.java
@@ -1,5 +1,5 @@
 /*
- * EditingPreferencesPane.java
+ * CodePreferencesPane.java
  *
  * Copyright (C) 2022 by Posit Software, PBC
  *
@@ -99,11 +99,13 @@ public class CodePreferencesPane extends PreferencesPane
       editingPanel.add(useNativePipe_);
       editingPanel.add(checkboxPref(constants_.editingReindentOnPasteLabel(), prefs_.reindentOnPaste()));
       editingPanel.add(checkboxPref(constants_.editingVerticallyAlignArgumentsIndentLabel(), prefs_.verticallyAlignArgumentsIndent()));
-      editingPanel.add(checkboxPref(prefs_.softWrapRFiles()));
       editingPanel.add(checkboxPref(
             constants_.editingContinueCommentsOnNewlineLabel(),
             prefs_.continueCommentsOnNewline(),
             constants_.editingContinueCommentsOnNewlineDesc()));
+      editingPanel.add(checkboxPref(
+            constants_.displayEnableTextDragLabel(),
+            prefs_.enableTextDrag()));
       editingPanel.add(checkboxPref(
             constants_.editingHighlightWebLinkLabel(),
             prefs_.highlightWebLink(),
@@ -216,18 +218,13 @@ public class CodePreferencesPane extends PreferencesPane
       displayPanel.add(checkboxPref(constants_.displayRelativeLineNumbersLabel(), prefs_.relativeLineNumbers()));
       displayPanel.add(tight(showMargin_ = checkboxPref(constants_.displayShowMarginLabel(), prefs_.showMargin(), false /*defaultSpace*/)));
       displayPanel.add(indent(marginCol_ = numericPref(prefs_.marginColumn())));
+      displayPanel.add(checkboxPref(prefs_.softWrapRFiles()));
+      displayPanel.add(checkboxPref(prefs_.marginColumnSoftWrap()));
+      displayPanel.add(checkboxPref(prefs_.marginColumnEditorWidth()));
       displayPanel.add(checkboxPref(constants_.displayShowInvisiblesLabel(), prefs_.showInvisibles()));
-      indentGuides_ = new SelectWidget(
-         prefs_.indentGuides().getTitle() + ":",
-         (Prefs.EnumValue) prefs_.indentGuides(),
-         false,
-         true,
-         false);
-      displayPanel.add(indentGuides_);
       displayPanel.add(checkboxPref(constants_.displayBlinkingCursorLabel(), prefs_.blinkingCursor()));
       displayPanel.add(checkboxPref(constants_.displayScrollPastEndOfDocumentLabel(), prefs_.scrollPastEndOfDocument()));
-      displayPanel.add(checkboxPref(constants_.displayEnableTextDragLabel(), prefs_.enableTextDrag()));
-      
+
       foldMode_ = new SelectWidget(
             constants_.displayFoldStyleLabel(),
             (Prefs.EnumValue) prefs_.foldStyle(),
@@ -236,6 +233,15 @@ public class CodePreferencesPane extends PreferencesPane
             false);
 
       displayPanel.add(foldMode_);
+
+      indentGuides_ = new SelectWidget(
+         prefs_.indentGuides().getTitle() + ":",
+         (Prefs.EnumValue) prefs_.indentGuides(),
+         false,
+         true,
+         false);
+
+      displayPanel.add(indentGuides_);
 
       displayPanel.add(headerLabel(constants_.syntaxHeaderLabel()));
       displayPanel.add(checkboxPref(prefs_.highlightRFunctionCalls()));