+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTargetPrefsHelper.java
@@ -111,6 +111,19 @@ public class TextEditingTargetPrefsHelper
             {
                docDisplay.setPrintMarginColumn(arg);
             }));
+      releaseOnDismiss.add(prefs.marginColumnSoftWrap().bind(
+            (arg) ->
+            {
+               if (arg)
+               {
+                  int column = prefs.marginColumn().getValue();
+                  docDisplay.setWrapLimitRange(column, column);
+               }
+               else
+               {
+                  docDisplay.setWrapLimitRange(null, null);
+               }
+            }));
       releaseOnDismiss.add(prefs.showInvisibles().bind(
             (arg) ->
             {